+++ crates/gui/CHANGELOG.md
@@ -6,7 +6,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/Gui-Yom/hlbc/compare/gui-v0.2.1...HEAD)
+## [Unreleased](https://github.com/Gui-Yom/hlbc/compare/gui-v0.3.0...HEAD)
+
+## [0.3.0](https://github.com/Gui-Yom/hlbc/compare/gui-v0.2.1...gui-v0.3.0) - 2023-11-16
 
 ### Added